requirements.txt
Flask==2.3.3
lmstudio
PyPDF2==3.0.1
python-docx==1.1.0
textract==1.6.3